트랜잭션, 구글API빼고 함

+++ src/main/java/com/example/project_team/dto/LoanDTO.java
@@ -25,5 +25,5 @@ public class LoanDTO {
 	private int minPeriod;
 	private int maxPeriod;
 	private String repayment;
-	private int commission;
+	private double commission;
 }
